add knobs
