+++ src/client/components/Description Career/DescriptionCareer.js
@@ -13,26 +13,9 @@ function DescriptionCareer({ title, content, content2, time }) {
   );
 }
 
-const arr = [
-  {
-    title: 'CAREER PATH ',
-    content:
-      'Curious about the possibilities within the IT-industry and want to know how to fit in job market?',
-    content2: 'Then try out the quiz and see the answers.',
-    duration: 3,
-  },
-  {
-    title: 'PRESENT YOURSELF',
-    content:
-      'Interested in tips and advices that will help you grow professionally?',
-    content2: 'Then climb up those professional stairs with our elevator.',
-    duration: 5,
-  },
-];
-
-export const Description = () => (
+export const Description = ({ exams = [] }) => (
   <div className="innerDiv">
-    {arr.map((item) => (
+    {exams.map((item) => (
       <div key={item.duration}>
         <DescriptionCareer
           title={item.title}
